VRT: fix warnings introduced by previous commit
git-svn-id: https://svn.osgeo.org/gdal/trunk@32758 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtrawrasterband.cpp
@@ -326,7 +326,7 @@ CPLErr VRTRawRasterBand::XMLInit( CPLXMLNode * psTree,
 
     const char* pszImageOffset = CPLGetXMLValue( psTree, "ImageOffset", "0");
     const vsi_l_offset nImageOffset = CPLScanUIntBig(
-                                    pszImageOffset, strlen(pszImageOffset));
+                        pszImageOffset, static_cast<int>(strlen(pszImageOffset)));
 
     if( CPLGetXMLValue( psTree, "PixelOffset", NULL ) == NULL )
         nPixelOffset = nWordDataSize;
